+++ esp-radio/src/ieee802154/pib.rs
@@ -101,7 +101,7 @@ pub(crate) fn ieee802154_pib_init() {
             coordinator: false,
             promiscuous: true,
             rx_when_idle: false,
-            txpower: 10,
+            txpower: 20,
             channel: 11,
             pending_mode: PendingMode::Disable,
             multipan_mask: 1 << IEEE802154_MULTIPAN_0,
@@ -227,7 +227,7 @@ fn ieee802154_txpower_convert(txpower: i8) -> u8 {
             const IEEE802154_TXPOWER_INDEX_MIN: i8 = 3;
         }
     }
-    if txpower > IEEE802154_TXPOWER_VALUE_MAX {
+    if txpower >= IEEE802154_TXPOWER_VALUE_MAX {
         15
     } else if txpower <= IEEE802154_TXPOWER_VALUE_MIN {
         IEEE802154_TXPOWER_INDEX_MIN as u8